+++ NEWS.md
@@ -1,6 +1,20 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
+# irtQ 0.1.1
+
+- Resolved the misalignment issue of standard errors in the output of
+  the `est_irt()` function when `fix.a.1pl = TRUE` is specified and the
+  items are calibrated using the 1PLM.
+
+- Added a new function, `grdif()`, to perform differential item
+  functioning (DIF) analysis across multiple groups. This function
+  calculates three generalized IRT residual DIF (GRDIF) statistics. For
+  more information about the function and its usage, please refer to the
+  accompanying documentation.
+
+- Fixed several typos in the manual documentation
+
 # irtQ 0.1.0
 
 - Initial release on CRAN